requirements.txt
geopy==2.3.0
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
python-telegram-bot-calendar==1.0.5
pytz==2022.7
requests==2.28.1
schedule==1.1.0
SQLAlchemy==1.4.44
timezonefinder==6.1.9
